+++ src/misha-init.cpp
@@ -40,8 +40,8 @@ extern "C" {
     extern SEXP gintervsort(SEXP, SEXP);
     extern SEXP gintervunion(SEXP, SEXP, SEXP);
     extern SEXP giterator_intervals(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
-    extern SEXP gmapply(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
-    extern SEXP gmapply_multitask(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
+    extern SEXP gmapply(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
+    extern SEXP gmapply_multitask(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
     extern SEXP C_gpartition(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
     extern SEXP C_gquantiles(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
     extern SEXP gquantiles_multitask(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
@@ -112,8 +112,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"gintervsort", (DL_FUNC)&gintervsort, 2},
     {"gintervunion", (DL_FUNC)&gintervunion, 3},
     {"giterator_intervals", (DL_FUNC)&giterator_intervals, 6},
-    {"gmapply", (DL_FUNC)&gmapply, 8},
-    {"gmapply_multitask", (DL_FUNC)&gmapply_multitask, 8},
+    {"gmapply", (DL_FUNC)&gmapply, 9},
+    {"gmapply_multitask", (DL_FUNC)&gmapply_multitask, 9},
     {"C_gpartition", (DL_FUNC)&C_gpartition, 8},
     {"C_gquantiles", (DL_FUNC)&C_gquantiles, 6},
     {"gquantiles_multitask", (DL_FUNC)&gquantiles_multitask, 6},